Authentication + Authorization

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Post;
 use App\Like;
 use App\Tag;
+use Auth;
+use Gate;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
@@ -53,11 +55,12 @@ class PostController extends Controller
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
+        $user = Auth::user();
         $post = new Post([
             'title' => $request->input('title'),
             'content' => $request->input('content')
         ]);
-        $post->save();
+        $user->posts()->save($post);
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
     //    $post->addPost($session, $request->input('title'), $request->input('content'));
         return redirect()->route('admin.index')->with('info', 'Post created, Title is: ' . $request->input('title'));
@@ -70,6 +73,10 @@ class PostController extends Controller
             'content' => 'required|min:10'
         ]);
         $post = Post::find($request->input('id'));
+        /*if(Gate::denies('manipulate-post', $post)){
+            return redirect()->back();
+        }
+        */
         $post->title=$request->input('title');
         $post->content=$request->input('content');
         $post->save();
@@ -77,9 +84,13 @@ class PostController extends Controller
         return redirect()->route('admin.index')->with('info', 'Post edited, new Title is: ' . $request->input('title'));
     }
 
-    public function getAdminDelete($id)
-    {
+    public function getAdminDelete($id){
         $post = Post::find($id);
+      /*
+         if(Gate::denies('manipulate-post', $post)){
+            return redirect()->back();
+        }
+        */
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
